fix: Rename ourlog -> log
