+++ packages/core/src/types-hoist/log.ts
